+++ src/main/java/frc/robot/commands/shooter/Shoot.java
@@ -1,26 +1,25 @@
 package frc.robot.commands.shooter;
 
 import edu.wpi.first.wpilibj2.command.Command;
-import frc.robot.subsystems.intake.Intake;
+import frc.robot.subsystems.indexer.Indexer;
 
 public class Shoot extends Command {
-  private final Intake m_intake;
+  private final Indexer m_indexer;
   private boolean m_reverse;
 
-  public Shoot(Intake intake, boolean reverse) {
-    m_reverse = reverse;
-    m_intake = intake;
+  public Shoot(Indexer indexer, boolean reverse) {
+    m_indexer = indexer;
 
-    addRequirements(m_intake);
+    addRequirements(m_indexer);
   }
 
   @Override
   public void initialize() {
-    m_intake.startFeedNote(m_reverse);
+    m_indexer.startFeedNote(m_reverse);
   }
 
   @Override
   public void end(boolean interrupted) {
-    m_intake.stopFeedNote();
+    m_indexer.stopFeedNote();
   }
 }
